Add handlers for all undecorated stage resizing events
